added more assertions in logs test and moved them to console
